Added footnotes styling and some configuration
Moved file locations around to keep SASS sources out of static directory
and also used SASS more effectively to make it easier to maintain the
two output files for the two current colour schemes. Also numerous
tweaks, such as a configurable copyright message and a heading prefix
string (for which I use a unicode "tick" symbol), as well as styling for
Markdown footnotes.