Added breadcrumb trail and tag cloud support
This commit removes the hard-coded "andy-pearce.com" from the upper and
lower left corners of the base template, which must have slipped through
the net when I originally removed site-specific hacks. In its place
there's now a simple breadcrumb trail providing the context of the
current page. This commit also includes updates to support generation of
a tag cloud, with up to 5 size steps (the `TAG_CLOUD_STEPS` setting in
Pelican).