Added archive & page support, code block styling
This commit includes many changes. The major changes are that it adds
support for archives and static pages to the template, so these features
should now work gracefully and have entries added to the menu if
requested. Also, support for a "tweet" button in the article header is
added, if `TWITTER_USERNAME` is defined. There have also been some other
minor tweaks to the templates for stylistic reasons and removal of
legacy code which was erroneously cloned from the original template.

On the CSS front, the monospace font was changed from Lekton to Ubuntu
Mono in an attempt to fix an issue where the line number column become
misaligned with the main `<pre>` block in some browsers when bold and/or
italics were used. The change to Ubuntu Mono only partially fixed the
issue, so all bold and italic Pygments styles have now been removed.
Debatably it's clearer to avoid these in code blocks anyway. The change
to Ubuntu Mono remains as it seems to be clearer to read in some cases.

Also on the Pygments front, the `pygments.css` file has been migrated to
SASS to reduce manual duplication of colours. Also, the command-line
parameters to the SASS compiler have been changed to render compressed
output, in an attempt to reduce page loading times. I suspect that other
factors (such as the use of Google Fonts) are the dominant constraints
on loading time, but this is an easy change and may help a little.