Update Google Analytics and heading styles
This commit updates the template for Google Analytics to reflect the
snippet currently presented by their service for universal tracking.
This requires the domain name as well as the key, so an extra variable
is required in the configuration file.

This commit also tidies up the style for sub-headings within a section.