Add multi-part support, add summary to all posts
This commit includes two changes. Firstly, the multi-part extension is
now supported, including an "other posts in this series" section if more
than one post is added with the same part name - this currently only
displays on the article view.. Secondly the article summary is now
included as a small byline just under the post title on both index and
article views.