Improvements to article series and pagination
This commit tidies up the handling of article series such that the
entire list of articles is not reproduced on the recent articles page,
only on the individual article page. There are also some presentation
tweaks to pagination, layout and style. Finally there's also a
updatecss.py script for use on systems where sass isn't installed - for
example,this avoids installing Ruby (and all its dependencies) on
servers which wouldn't otherwise need it.