Added Solarized colour scheme
This commit adds a dark-on-light colour scheme based on the
[solarized palette](http://ethanschoonover.com/solarized)
created by Ethan Schoonover. This involves a little reworking, and also
has some minor impact on the existing schemes. In particular, <pre>
blocks no longer use a different shade of background or a border - this
was a change I'd been considering anyway, since I feel it makes the text
feel less choppy when interspersing lots of code snippets into text.