Fixed minor issues with article parts
I noticed on some platforms the two unicode characters U+2610 (BALLOT
BOX) and U+2611 (BALLOT BOX WITH CHECK) are different widths and
stylistically dissimilar. One example of this is iOS platforms, where
U+2610 is in the standard font but U+2611 is presented in the same
style as the Apple emoji.

I can't do a lot about the latter issue but I could fix the former by
some CSS tweaks - the :before elements for the <li> tags that make up
the parts list are now floats with a min-width and centered alignment. I
also added emphasis to the current article in the list.