Back to sans-serif font.
As pretty as I find serifs, I find it an inescapable conclusion that
they're simply less readable on screens. While this is less of an issue
than it used to be, due to the proliferation of retina displays, there
are still a good number of people reading these things on cheap monitors
where the manufacturer has cranked up the display size without
maintaining the DPI.

Hence, it's back to Alegreya for the body text.