Fix monospace font size
With the switch back to a sans-serif body font, monospaces now look too
large in running text (this is regardless of the recent change to Share
Tech Mono for the monospace font). Hence we reduce them to 90% size, but
only in running text - nothing wrong with code snippets looking larger
than running text, they need to be clearly presented.

I've also added a small margin to <tt> elements - this is more related
to the change to Share Tech Mono, with its wide character spacing, but
something I've been meaning to do for awhile anyway as I don't like it
when identifiers used in text merge visually with the non-identifier
text around them.