Added mobile styling
This is a first stab at an alternative sylesheet which is suitable for
smaller devices. This allows pages to be reactive and switch to this as
the window width reduces. This isn't necessarily the most graceful option
available on modern browsers, but it's a fairly low-hanging fruit and
doesn't rely on a sharp learning curve in new CSS. I can always improve
it later to use a somewhat less clumsy approach when I have time.