Mobile font size fixes
This commit removes the probably always rather daft idea of running the
text on the whole page larger than is configured in the browser. The
only argument for it is if the font size on Google Fonts isn't quite set
correctly, but it seems unlikely that my patchy testing is more reliable
than that (especially as I only just realised I had my Google Chrome
font size set to "small" for some reason!).

There are also some small fixes to margins and other sizing issues, and
I've also disabled text size adjustment on the mobile stylesheet. Since
my changes there shouldn't be any need for it, since this is just a
full-width style and the browser is free to lay it out however it likes.
I believe I've removed all pixel sizes from the layout - if I haven't,
I'll address in later commits. AFAICT these are the only justification
for the text size adjustment in the first place.