Impose max-width on printed text
Readability is degraded if text is too wide, so I'm imposing the same
max-width: 45em on the print stylesheet than I do with the screen one.
This is probably still slightly too wide, but if I constrain it too much
then the <pre> blocks are likely to overflow.

Left-aligned text that doesn't fill the width of the page looks odd when
printed so I've also made it centered - this is a lot easier in the
print stylesheet where there's no sidebar to worry about.