Updated footnote Javascript
This commit switches to an alternative Javascript engine for the
footnote prettification - instead of popping up the footnote on hover,
this one pops up a window at the base of the screen on click. This is
more useful on mobile devices.

I did consider removing the footnotes panel at the base of the article
entirely, but two things gave me pause: firstly I want to remove it only
for screen media AND if Javascript is supported and enabled in the
browser - this is quite hard with stanadard media queries; secondly I
don't want to harm the experience on very basic devices. There's
something nice about the notes being sat right there even in case of
some catastrophic Javascript failure.

This commit also removes the left padding on section divs which was
causing the dividing lines to no longer span the whole element. As and
when I find elements that were benefitting from it, I'll have to add
padding or margin to those on a case-by-case basis.

I made a few tweaks to make it work with Pelican, but all credit to the
original author:

https://github.com/andytlr/jquery-multimarkdown-footnotes