Capped size of footnote popup
Since I limit the width of the text in my main sections it seemed inconsistent
to then fail to limit the width of my footnote popups -- on a large window the
fixed 100% width gets substantially harder to read.

Instead of limiting just the text within it, I capped the width of the whole
popup and also right-aligned in -- this allows a user with a large display
to minimise the chances of the popup overlapping with body text by putting it
in the dead space otherwise left to the right.