Stylistic improvements to blockquotes
This commit adds quote marks at the top-left and bottom-right corners to
the blockquote block. They're the same colour as the section background,
partly for stylistic reasons and partly because if something goes wrong
and they extend outside the boundary of the box, they won't show up.

I'm not 100% convinced by how robust the CSS will be in several
environments and browsers, but it seems to work in Safari and Chrome so
I think I'm pretty well covered. Nobody uses IE any more, do they?