Various small changes to margins, footnotes and series summaries
This commit makes assorted small changes to margins for consistency and
appearance on mobile and other layouts. It also adds square brackets
around references for footnotes, since often those tiny subscripted
numbers were very easy to miss. It also duplicates the series list at
the bottom when viewing a single article since there's a reasonable
chance that the viewer will want to jump straight to the next article in
the list and it seems irritating to force them to scroll all the way to
the top for that.