Move to JS site menu
The hover menu with CSS is suboptimal on touch devices, so I'm moving to
a simple JS implementation. I've also moved it on to the left bar to
replace the barely used "Links" section, so there's a fallback option
for any site without JS. Finally, this commit fixes some issues
introduced with python-markdown-math awhile ago.