Fixed layout on intermediate size viewports
For some reason `div.document` had `display: table` set on it, which was
causing it to become larger than the viewport on sizes which were
smaller than the `max-width` but larger than the pixel width required to
trigger the mobile stylesheet.

The fix was just to remove `display: table`, which didn't seem to be
doing anything for us and was only on the non-mobile stylesheet. This
allows the `div.section` width to fluidly shrink as the viewpoint
resizes up to the point where the width goes over the threshold for the
mobile stylesheet.