Fix <dl> styling
I finally worked out that the odd bullet sizing for <dl> lists was due
to the bold weight being used for <dt> items. Quite why bold weight
makes the bullets small I'm not sure, but this seems to fix the issue on
every browser I tried.

I also updated the styling to add a left border for the <dd> items,
which seems more consistent with the rest of the look.