Table support + new monospace font
This commit includes support for tables, as well as a new monospace font and a
change of colour of monospace text to the highlight colour (currently blue), as
well as a new monospace font which is my latest effort in my search for a font
which is clear but still stands out from non-monospace text. There are also a
slew of other visual changes, such as adding dotted underlines on headings.